drop功能重构

+++ ly-kernel/src/name/huliqing/core/object/el/LevelEl.java
@@ -22,11 +22,16 @@ public class LevelEl<T extends ElData> extends AbstractEl<T> {
     
     // 因为表达式不会在运行过程中发生变化，所以对于levelEl来说是可以进行缓存的。
     // 缓存的最高数量不会超过系统定义的最高等级数
-    private final Map<Integer, Double> cacheMap = new HashMap<Integer, Double>();
+    private final Map<Integer, Float> cacheMap = new HashMap<Integer, Float>();
     
-    public synchronized double getValue(int level) {
-        // 先从缓存中获取。
-        Double result = cacheMap.get(level);
+    /**
+     * 获取等级值，如果找不到指定的等级值，则该方法始终返回0.
+     * @param level
+     * @return 
+     */
+    public synchronized float getValue(int level) {
+        // 先从缓存中获取
+        Float result = cacheMap.get(level);
         if (result != null) {
             return result;
         }
@@ -36,9 +41,9 @@ public class LevelEl<T extends ElData> extends AbstractEl<T> {
         if (strResult != null) {
             try {
                 
-                result = Double.parseDouble(strResult);
+                result = Float.parseFloat(strResult);
                 cacheMap.put(level, result);
-                return result.doubleValue();
+                return result;
                 
             } catch (NumberFormatException nfe) {
                 Logger.getLogger(LevelEl.class.getName())
